refactor/#88 : 패키지 autoadd/directadd로 구분 및 호이스팅 변경
