+++ app/src/main/java/com/kiero/presentation/parent/navigation/ParentNavigation.kt
@@ -9,8 +9,8 @@ import androidx.navigation.compose.navigation
 import com.kiero.core.navigation.Route
 import com.kiero.presentation.parent.screen.alarm.navigation.parentAlarmNavGraph
 import com.kiero.presentation.parent.screen.journey.navigation.parentJourneyNavGraph
-import com.kiero.presentation.parent.screen.mission.auto.navigation.parentAutoMissionAddNavGraph
-import com.kiero.presentation.parent.screen.mission.navigation.parentMissionAddNavGraph
+import com.kiero.presentation.parent.screen.mission.autoadd.navigation.parentAutoMissionAddNavGraph
+import com.kiero.presentation.parent.screen.mission.directadd.navigation.parentMissionAddNavGraph
 import com.kiero.presentation.parent.screen.mypage.navigation.parentMypageNavGraph
 import com.kiero.presentation.parent.screen.reward.navigation.parentRewardNavGraph
 import com.kiero.presentation.parent.screen.schedule.navigation.parentScheduleNavGraph
