--- app/src/main/java/com/kiero/presentation/parent/screen/mission/auto/ParentAutoLoadingScreen.kt
@@ -1,82 +1,0 @@
-package com.kiero.presentation.parent.screen.mission.auto
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.kiero.R
-import com.kiero.core.designsystem.component.KieroGifImage
-import com.kiero.core.designsystem.component.KieroSnackbar
-import com.kiero.core.designsystem.theme.KieroTheme
-
-@Composable
-fun ParentAutoLoadingScreen(
-    paddingValues: PaddingValues,
-    snackbarHostState: SnackbarHostState,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier
-            .fillMaxSize()
-            .background(color = KieroTheme.colors.black)
-            .padding(paddingValues)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .offset(y = (-40).dp),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            KieroGifImage(
-                drawableId = R.drawable.gif_parent_analyze
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            Text(
-                text = "알림장을 분석하고 있어요!",
-                style = KieroTheme.typography.semiBold.title2,
-                color = KieroTheme.colors.white
-            )
-        }
-
-        SnackbarHost(
-            hostState = snackbarHostState,
-            modifier = Modifier
-                .align(Alignment.BottomCenter)
-                .padding(bottom = 65.dp)
-        ) { data ->
-            KieroSnackbar(
-                message = data.visuals.message,
-                modifier = Modifier.padding(horizontal = 16.dp)
-            )
-        }
-    }
-}
-
-@Preview
-@Composable
-private fun ParentAutoLoadingScreenPreview() {
-    KieroTheme {
-        ParentAutoLoadingScreen(
-            paddingValues = PaddingValues(),
-            snackbarHostState = remember { SnackbarHostState() }
-        )
-    }
-}
