+++ app/src/main/java/com/kiero/presentation/parent/screen/mission/autoadd/state/AutoMissionSideEffect.kt
@@ -1,0 +1,8 @@
+package com.kiero.presentation.parent.screen.mission.autoadd.state
+
+sealed interface AutoMissionSideEffect {
+    data class ShowToast(val message: String) : AutoMissionSideEffect
+    data class ShowToastAndNavigate(val message: String) : AutoMissionSideEffect
+    data object NavigateBack : AutoMissionSideEffect
+    data class ScrollToPage(val index: Int) : AutoMissionSideEffect
+}
